feat: Fetch community rehoming shelter ID
Refactors the helper function `getOrCreatePersonalShelter` to `getCommunityShelterId`. This change simplifies the logic by directly fetching the ID of a pre-defined "Dumble's Door Community Rehoming" shelter.

This assumes that a shelter with the name "Dumble's Door Community Rehoming" already exists in the `shelters` table. If it doesn't, an error will be thrown, indicating a configuration issue.